requirements.txt
flask==2.0.1
pytesseract==0.3.8
Pillow==9.0.0
PyPDF2==3.0.0
python-docx==0.8.11
python-pptx==0.6.21
pandas==1.3.3
spacy==3.1.3
transformers==4.11.3
torch==1.9.1
werkzeug==2.0.1
flask-cors==4.0.0
